docker/requirements.txt
pandas
requests
azure-eventhub